Added checking input for duplicates
