+++ push_swap.c
@@ -6,38 +6,11 @@
 /*   By: rnovotny <rnovotny@student.42prague.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/17 20:39:52 by rnovotny          #+#    #+#             */
-/*   Updated: 2023/10/12 17:34:34 by rnovotny         ###   ########.fr       */
+/*   Updated: 2023/10/12 21:33:14 by rnovotny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
-
-#include <stdio.h>
-
-long	ft_atoi(const char *nptr)
-{
-	int		i;
-	int		neg;
-	long	res;
-
-	neg = 1;
-	res = 0;
-	i = 0;
-	while ((nptr[i] >= 9 && nptr[i] <= 13) || nptr[i] == 32)
-		i++;
-	if (nptr[i] == '-' || nptr[i] == '+')
-	{
-		if (nptr[i] == '-')
-			neg = -neg;
-		i++;
-	}
-	while (nptr[i] >= 48 && nptr[i] <= 57)
-	{
-		res = res * 10 + (nptr[i] - 48);
-		i++;
-	}
-	return (neg * res);
-}
 
 void	sort_stacks(int **stacks, unsigned short *operations)
 {
@@ -55,7 +28,7 @@ void	sort_stacks(int **stacks, unsigned short *operations)
 	if (stacks[0][0] < 6)
 		special_cases(stacks, operations);
 	else
-		turk_sort(stacks, operations);
+		start_sort(stacks, operations);
 }
 
 void	process_stack(int argc, char **argv)
@@ -77,13 +50,31 @@ void	process_stack(int argc, char **argv)
 		stacks[0][i] = ft_atoi(argv[i]);
 	sort_stacks(stacks, operations);
 	read_operations(operations);	// post-optimisation
-	printf("process_stack, final printing:\n");
-	print_stacks(stacks);	// final check, do not forge to remove this!
 	i = -1;
 	while (i++ < 4)
 		free(stacks[i]);
 	free(stacks);
 	free(operations);
+}
+
+int	check_duplicates(int length, char **input)
+{
+	int	i;
+	int	j;
+	int	number;
+
+	i = 0;
+	while (++i < length)
+	{
+		number = ft_atoi(input[i]);
+		j = i;
+		while (++j < length)
+		{
+			if (number == ft_atoi(input[j]))
+				return (1);
+		}
+	}
+	return (0);
 }
 
 int	check_input(int length, char **input)
@@ -122,7 +113,7 @@ int	main(int argc, char **argv)
 		return (1);
 	if (check_input(argc, argv))
 	{
-		write(1, "Error\n", 6);
+		write(2, "Error\n", 6);
 		return (1);
 	}
 	i = 0;
@@ -130,9 +121,14 @@ int	main(int argc, char **argv)
 	{
 		if (ft_atoi(argv[i]) < -2147483648 || ft_atoi(argv[i]) > 2147483647)
 		{
-			write(1, "Error\n", 6);
+			write(2, "Error\n", 6);
 			return (1);
 		}
+	}
+	if (check_duplicates(argc, argv))
+	{
+		write(2, "Error\n", 6);
+		return (1);
 	}
 	process_stack(argc, argv);
 	return (0);
