+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: rnovotny <rnovotny@student.42prague.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/17 20:38:50 by rnovotny          #+#    #+#             */
-/*   Updated: 2023/10/12 17:38:40 by rnovotny         ###   ########.fr       */
+/*   Updated: 2023/10/12 21:46:11 by rnovotny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,9 +18,9 @@
 
 int		main(int argc, char **argv);
 int		check_input(int length, char **input);
+int		check_duplicates(int length, char**input);
 void	process_stack(int argc, char **argv);
 void	sort_stacks(int **stacks, unsigned short *operations);
-long	ft_atoi(const char *nptr);
 
 void	swap(int **stacks, int to_swap, unsigned short *operations);
 void	push_a(int **stacks, unsigned short *operations);
@@ -28,17 +28,21 @@ void	push_b(int **stacks, unsigned short *operations);
 void	rotate(int **stacks, int to_rotate, unsigned short *operations);
 void	reverse_rotate(int **stacks, int to_rotate, unsigned short *operations);
 
+int		find_min_index(int **stacks);
+int		find_min_max(int *stack, int min_or_max);
+int		find_offset(int **stacks);
+long	ft_atoi(const char *nptr);
+
+void	do_sort(int **stacks, unsigned short *operations);
+void	start_sort(int **stacks, unsigned short *operations);
 void	assign_score(int **stacks, int position);
 void	calculate_score(int **stacks, int a, int b);
-int		find_min_max(int *stack, int min_or_max);
-int		find_min_index(int **stacks);
-int		find_offset(int **stacks);
+void	calculate_score_rrr(int **stacks, int a, int b);
 
+void	move_rr(int **stacks, unsigned short *operations);
+void	move_rrr(int **stacks, unsigned short *operations);
+void	move_ra_rb(int **stacks, unsigned short *operations);
 void	special_cases(int **stacks, unsigned short *operations);
-void	turk_sort(int **stacks, unsigned short *operations);
-void	move(int **stacks, unsigned short *operations);
-
-void	print_stacks(int **stacks);
 
 void	two(int **stacks, unsigned short *operations);
 void	three(int **stacks, unsigned short *operations);
